feat(operators): Adiciona aprendizado com operadores aritmeticos e de atribuicao

+++ 04-projects/01-bootcamps/01-bradesco/01-basic-syntax/02-variables/src/Main.java
@@ -18,6 +18,6 @@ public class Main {
 
         //DECLARACAO DE VARIAVEL CONSTANTE
         //SEMPRE ESCREVER EM CAIXA ALTA
-        final int PI = 3.14;
+        final double PI = 3.14;
     }
 }
